Add next auth client session authentication
