--- models/authSchema.js
@@ -1,13 +1,0 @@
-const mongoose = require('mongoose');
-
-const userSchema = new mongoose.Schema({
-  username: String,
-  firstName: String,
-  lastName: String,
-  email: String,
-  password: String
-});
-
-const User = mongoose.model('User', userSchema);
-
-module.exports = User;
